pyproject.toml
[project]
name = "meeting-scribe"
version = "0.1.0"
dependencies = [
    "django",
    "djangorestframework",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "azure-storage-blob (>=12.25.1,<13.0.0)",
    "mutagen (>=1.45.1,<2.0.0)",
]
